исправил ошибку при входе. добавил реализацию добавления классности. убрал отображение текста ошибок в удалении/добавлении записей при переключении между справочниками или табами просмотра/редактирования

+++ src/Staffinfo.Desktop/Model/ClasinessModel.cs
@@ -19,7 +19,7 @@ namespace Staffinfo.Desktop.Model
         /// Номер приказа
         /// </summary>
         [DisplayName("Номер приказа")]
-        public ushort OrderNumber { get; set; }
+        public int OrderNumber { get; set; }
 
         /// <summary>
         /// Дата подтверждения классности
@@ -31,7 +31,7 @@ namespace Staffinfo.Desktop.Model
         /// Уровень классности
         /// </summary>
         [DisplayName("Уровень классности")]
-        public byte ClasinessLevel { get; set; }
+        public int ClasinessLevel { get; set; }
 
         /// <summary>
         /// Описание
